+++ process/rating/export_test.go
@@ -1,5 +1,5 @@
 package rating
 
-func (p *BlockSigningRater) SetRatings(newRatings map[string]int64) {
+func (p *BlockSigningRater) SetRatings(newRatings map[string]uint32) {
 	p.ratings = newRatings
 }
